automation/kronos/requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2